requirements.txt
# Core dependencies
click>=8.0.0
apify-client>=1.6.0
openai>=1.0.0
yt-dlp>=2024.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Optional: YouTube upload
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
